Move osmomath into its own go.mod
