+++ go.work
@@ -1,0 +1,4 @@
+go 1.18
+
+use ./osmomath
+use .
