+++ x/gamm/pool-models/stableswap/amm_bench_test.go
@@ -4,7 +4,7 @@ import (
 	"math/rand"
 	"testing"
 
-	"github.com/osmosis-labs/osmosis/v13/osmomath"
+	"github.com/osmosis-labs/osmosis/osmomath"
 )
 
 func BenchmarkCFMM(b *testing.B) {
